Moved requests to request file and moved upload requests from client/orgs requests
